Changes made
Changes Made:
Dropdown for Role Selection in User Edit:

When editing a user, the role field is now a dropdown <select>, allowing the admin to choose between User, Admin, and Super Admin.
The current role of the user is selected by default in the dropdown.
Add User Form:

The form for adding users still includes the username, password, and a dropdown for selecting the role.
Edit User:

You can update the username and role using text fields and a dropdown.
The password field is optional. If you want to update the password, enter a new password. If you leave it blank, the password will not be updated.
How it works:
Add User: Allows the admin to add a new user with a username, password, and role.
Edit User: Lets the admin edit the username and role of any user, with an optional password update.
Delete User: Allows the admin to delete users from the system.